Reverting Mart's bots 2.0, which was a gutting of Artur's bots 1.0, with my bots 3.0.  Here's hoping we can all get along now and:
-- Mart will have his nice theory way that no clients support
-- Brad and Artur will have nice configuation w/o ugly class decls for bot logic
-- Artur won't have callbacks,
-- Mart won't have duplicated logic code in components vs. roster items,
etc
etc

Basically I saw what Mart wanted (no duplicated code) and Artur wanted
(pragmatic), and I found a way to support both.

In new way, the bot is generic (like Mart wanted), but it's the 1st
class thing, not a component.  A component is just a bridge to it, in
the same way a rosteritem and messages to it are a bridge.

The roster item / presence / messages bridge is built.  Mart can build
the component one.

Also introducing a BotContext, which is what you reply with (and
eliminates need for callbacks, since it basically is one).  Read notes
at top of context... eventually we'll also have history and context
timeouts in there, for smarter bots.  For now it's pretty dumb, but
should lead the way.




git-svn-id: http://code.sixapart.com/svn/djabberd/trunk@586 43dd9337-660f-0410-bd32-e7601923a1a1