Completely rewritten implementation of DJabberd::Component::Easier for a nicer API and less code duplication.
Most of the fun stuff is now in DJabberd::Agent, which represents "thing to collect together methods that are used both by components and nodes". It's sick, but it was the best way I could think of to arrange these things to keep all of the common stanza processing stuff in one place.

git-svn-id: http://code.sixapart.com/svn/djabberd/trunk@625 43dd9337-660f-0410-bd32-e7601923a1a1